+++ src/main/kotlin/com/andreikingsley/domain/Payment.kt
@@ -1,6 +1,7 @@
 package com.andreikingsley.domain
 
 import jakarta.persistence.*
+import java.math.BigDecimal
 
 @Entity
 @Table(name = "payments")
@@ -24,5 +25,5 @@ class Payment(
     var paymentType: String,
 
     @Column(name = "payment_value")
-    var paymentValue: Double,
+    var paymentValue: BigDecimal,
 )
